Refactor update commands for clarity and consistency
